Add empty description error to Google style raises
I attempted to add the empty description error to the Numpy
style docstring's argument section.  However, it turns out that
without a description, the item name can be elided up unto the
word which represents the argument.  Since terminal productions
don't yet have annotations, the annotations are dropped.

Possible solutions to this will either be to allow annotations on
terminal productions, or annotate children of a production rather
than the production itself.