Add better analysis for raises.
Previously, we just looked for raise statements and grabbed
any name which was present and used that to match against the
docstring.  However, that's not a very good strategy.
For example, if you catch and throw (admittedly, probably a bad
pattern), then it would fail.

This commit adds a subclass of ast.NodeVisitor, RaiseVisitor.
It walks through the tree and maintains a count of raise statements
and their context.  It also maintains a partial symbol table for
those except statements.

The rules for this are somewhat obvious, as they follow from
Python.  A context is introduced for the function/method.  Then,
for each try block, a context is introduced.  The except handlers
are analyzed, and raised exceptions are removed if they were
handled.  If a name was present in the exception handler, then
it adds it to the symbol table and points to the exception class.

This is not at all complete, and will certainly break in a lot
of circumstances.  (For example, if you store an exception in
a variable in a block, and then raise that.)  For that reason,
when it breaks, it's a soft break: darglint will display an
error.  Maybe I'll get to implementing a more thorough implementation
at some point.  Probably not.  Until then, users should just
add a noqa statement.

Resolves: Issue #68.