Allow for missing short description in Sphinx.
Docstrings which dropped the short description and started with
a section had confusing error messages, as they would parse the
first section as a short description.  This change allows sections
to come first, if they'll parse correctly.

For the moment, I'm not adding an error for a missing short
description, as that would conflict with pydocstyle.  I may
change that in the future, as pydocstyle has started reporting
errors about the docstring's contents.