Move configuration to global variable.
Setting configuration as a global value significantly cleans
up the signatures of a lot of functions, and avoids a lot of
tramp data.

This will make it significantly easier to implement a custom
assert function, as these asserts are found throughout the codebase,
and it would be very messy to pass a configuration everywhere.

Resolves: Issue #85.