Exchange asserts for custom assert, add Path.
Now, most of the files use the custom assert for assertions,
which will make the program a little robust, and will help
with logging.

To add custom asserts to Identifiers, I've created a helper class
to describe and extract paths from parse trees.  This helper
class allows the path extraction to be strongly typed, and should
make walking the path a little safer.  This will also make the
Identifier subclasses entirely declarative -- all logic will happen
in the abstract Identifier class once it is integrated.