Fix race condition in integrity checker.
The docstring in the integrity checker was a member of
the class instance.  So when multiple runs happened, it
was possible for the member to get overriden.  This only
happened when two docstrings being evaluated had similar
parse times (otherwise, the variance is high enough that
all other checks went through.)

Resolves: Issue #83.