Handle fully/partially qualified caught exceptions.
Previously the below would have given an error in darglint:

    def finish_me():
        """A function to finish.

        Raises:
            requests.exceptions.ProxyError: Until
                this function is finished.

        """
        try:
            raise requests.exceptions.ProxyError()
        except requests.exceptions.ProxyError():
            raise

This was due to qualified names being represented differently
in the ast.  Darglint no longer reports this as an error.
However, the caught exception and the documented exception
must match exactly.  If you caught `ProxyError`, but documented
`requests.exceptions.ProxyError`, darglint will not recognize
them as being the same.

See: Issue #88