Fix error messages for underspecified types.
Previously, if you had a function like the following,

    def f(x: int, y: int, **kwargs):
        """Describes my function.

        Parameters
        ----------
        x, y: int
            The main parameters.
        kwargs
            The extra parameters

        """
        ...

It would result in confusing error messages. This is because darglint
grabs all the items (['x, y', 'kwargs']), then all of the types
(['int']), and then zips them to check them against the function
history (after expanding the arguments).  So we would end up with
[('x', 'int'), ('y', 'int')].  That means we would probably have
reported a missing argument, 'kwargs', even though it's specified.

The fix was to not zip items and types.  Instead, we extract all of
the item nodes from the AST, then we traverse each of those subtrees
and find all of the types, if any are present.

Resolves: Issue #112.