Allow users to set the log level.
The logger now has a configurable log level.  Each time
the configuration object sets the log level, it passes
that to the root logger, which gets passed to the darglint
logger.

To use, the flags `-l`, `--log-level` were added:

    darglint --log-level=INFO example.py

CRITICAL logs are to be reserved for displaying a message
before darglint exits by failure.  ERROR logs represent
some unexpected state (and should probably be reflected with
an assertion.)