An example reformatter.
This example is a proof of concept using darglint's existing API.
It uses a simple example (of formatting hexidecimal numbers) as an
example.

Formatting will have to occur in two steps:

1. Transform the tree to fix any complex errors which can be resolved.
2. Reconstruct the string, using the correct amount of whitespace, etc.

For the first step, it makes sense to have the fixing function
on the error.  Semantically, the error and the way to fix it are
closely related. (For example, the error message could give hints on how
to fix the error.)  This also ensures that we have the closes scope possible
when attempting to resolve an error.

The relevant details are in the grammar (which has to represent different
possible errors the user could make), the error (which has to attempt to
fix the error), the `fix_tree` function, which traverses the tree and fixes
it, and finally the `reconstruct_string` method on `CykNode`, which does
finishing touches that aren't explicitly errors.

See: Issue #42.