Add return type handling to return visitor.
For this, it was necessary to visit the `FunctionDef` and `AsyncFunctionDef`
of the tree, which means that the various mixins in the analysis directory
are no longer disjoint.

If this happens again, we'll have to find a better abstraction.  For the
moment, though, we can just call the super method in the `FunctionDef` and
`AsyncFunctionDef` visits for every class except the `FunctionScopedVisitorMixin`,
and make sure that we put that mixin first in the class definition for the
final visitor.

See: Issue 132.