Integrate all visitors into AnalysisVisitor.
Now that all of the visitors are complete, we can easily compose
them into a single visitor, the AnalysisVisitor.  This visitor
returns all of the attributes which should be part of the function
signature.  Some of the stored values are a little inconsistent
(sometimes I've stored strings, sometimes I've stored AST instances.)
We can make it more regular in the future.

Now all that's left is to integrate this into the function description,
and write some tests to check that nested functions are handled correctly.