Integrate analysis visitor into function description.
Most parts of the function description are found using the
AnalysisVisitor now.  The notable exceptions are those attributes
which can be fonud directly on the function nodes.

The next two steps to this refactoring are to use the
FunctionAndMethodVisitor to visit all of the functions and methods,
and create FunctionDescriptions from them.  Then, we should write
some classes which are capable of extracting human-readable
information from AST nodes. (For example, pulling the human-readable
information from a variable ast.Name by getting its `id` attribute.)
That should make it much cleaner overall.

I did have to remove the `FunctionDef` and `AsyncFunctionDef` visitors
from the `ReturnVisitor`.  It's not a big deal, since the function return
type can be extracted from those nodes directly.