Add makefile for dockerfiles and create docker tox section.
We can now run all of the supported python environments with docker through
the tox file.  This will make running tests locally a little bit easier
(assuming you have docker installed.)

I'm not sure why the docker images are so large -- around 300MB each.  I
imagine it's the installation of the dev dependencies and pip libraries.
I'll have to play with them and see if I can reduce their size in the
future.  For the moment, though, it doesn't matter that much.