Handle bare raise when exception captured in variable.
Previously, when we had a bare raise, we were gathering the exception
raised from context (or logging an error.)  For example, if we had
a function such as

    def f(x):
        try:
            1 / x
        except ZeroDivisionError:
            ... # Some processing
            raise

The exception would be correctly reported as being raised by the function,
and so a docstring describing `ZeroDivisionError` in its docstring
would be correct.  However, with a docstring such as

    def f(x):
        try:
            1 / x
        except ZeroDivisionError as err:
            ... # Some processing
            raise

It was not being reported as being raised by the function.  This was because
the variable was being stored separately from other exceptions in the context.
That way, if the user had a statement such as `raise err`, we would be able to
know what type of error it was.  However, I forgot to add this case to the
logic which handles the bare raise.

Resolves: Issue #134.