Fix the kfollow algorithm.
There were multiple issues with the kfollow algorithm, and resolving them
required a good deal of debugging.  But I think this is pretty close to
correct, now.

One problem was that when I was resolving a particular permutation of
production lengths, I wasn't allowing the last value to be taken from the
firstset of the last token (if it was a non-terminal.)  For example, if
we have the grammar

    S -> a S A
      | ε
    A -> a b S
      | ε

If we were trying to find the k = 2 followset of S, and we're examining the
production `S -> a S A`, then we have to take into consideration three
possible permutations.  That the terminal symbols which are produced by A are
of length 0, 1, and 2.  When considering the 0 permutation of A, we will still
have to find symbols to build up to k = 2, so we can only consider the fourth
production (the empty one.)  For the 1 permutation of A, we can only consider an
exact-length production of 1.  That's not possible, so we can skip it.  For a 2-
permutation of A, we can accept a partial production, because that will satisfy
our k = 2.  So we can use the firstset.  Previously, we were only using exact-length
productions, and so we were missing those.

However, that means that we can't append to that result when finding the fixpoint
solution.  (Because it's partial, and could lead to impossible productions.)  So,
we can only consider that when constructing completed followsets, which won't be
appended to.  (But which can occur on the later half of a partial followset.)

I'm unsure whether this is going to hold up for k > 2, and I think that the subproduction
produced from a firstset will have to store some state communicating that that's where
it comes from.  (Or, better yet, use a firstset directly.)