Added fuzz test for followsets.
This fuzz test generates random grammars, discarding any which have
infinite left recursion, and produces random languages from the grammar
in order to find a subset of all followsets.  It then calculates the actual
followsets, and makes sure that the actual followsets are a superset of
the first.

This is by no means a guarantee, but it does demonstrate that at we're on the
right track.  Of course, there could be followsets in the actual ones calculated
which are erroneous.  Manual inspection is probably the easiest way to verify
that there are not, so we can probably let it fail.