Clean up.
Factored classes into their own files, fixed some typing.  Added
black to the dev dependencies, and ran on overything.

Unfortunately, __getitem__ can take either an index of some type,
or a slice.  Then the return type could be different depending on
which one it takes.  There's no good way that I can find to convince
the type checker to perform the logic based on the calling type.
(If this were Java, we'd have method overloading, for example.)
So, I've just added some explicit type casts in places where this
is obvious.  In other places, I've used an assert. It's messy, but
oh well, I guess.