Clean up, add driver, add highlighter for CYK.
Mostly performed some cleaning of the generators file.  I've
moved all of the classes to their own files so that it's a little
easier to navigate and test.

I've added a failing test case, where the derivation to find isn't the
immediate result of a production.  Before handling that, I'll change
things to have it be more consistent.  For example, single symbols on the
RHS are now represented with a string, while more than one are represented
with a tuple.  I'll change them all to being tuples.  Then I'll see if
there is anything else I can clean up.  I'll remove any extra helpers for
seeing if something is a terminal/nonterminal/etc.

Once that's done, it will be a little easier to move forward.