Added pure abstract visitor and tests for it.
Added a visitor to determine if a function is pure abstract.
A function is considered abstract if it is decorated with @abstractmethod,
and considered pure if it has "no implementation".

Since Python does not have entirly empty functions, all functions which contain
one of the following (and possibly a docstring) are considered empty

 - ... (Ellipsis)
 - pass
 - raise NotImplementedException
 - return NotImplementedException
 - only a docstring