Ensure visitors from all mixins are called.
Ensure mixins use super().generic_visit() so that all mixins have a chance to
see each node. This is achieved through the base class VisitorBase which brings a
@VisitorBase.continue_visiting decorator.

Exceptions to this pattern are the function_scoped_visitor, since the whole point
of it would be undermind if the visitor descends further. For this reason the
function_scoped_visitor has to be the first base of the AnalysisVisitor.

Furthermore, the RaisesVisitor violates this policy, since its logic is implemented
in such a way, that it does not deal with descending deeply into nested try statements.
Doing so breaks the visitor. At this point I am not aware of another visitor using
visit_try or visit_raise, so I do not see a need to fix it right away.

An advantage of fixing this would be that VisitorBase could become a metaclass,
which automagically decorates all `visit_*` functions with @VisitorBase.continue_visiting.