Demonstration of issues with simpler callbacks.
Currently the added complexity of the visitor_base + decorator adds a lot of complexity.
Unfortunatly, I am afraid the complexity might be needed.
This is a minimal change showing how the simple calls
can potentially break things. This commit simply exists to show the reasoning behind the
complexity. However, in most parts of the mixins this is not needed,
thus an attempt to remove it where ever possible is made. This might serve as a reference
for the future, should changes start breaking.