Remove error logs when exceptions are from calls.
Since generating a class from a call can mean that any exception
can be handled, we should effictively just ignore it.  Users are
doing something weird, and they should expect to have to use
"noqa"s in their docstring.  In particular, because it's not always
going to be possible to know, statically, what exceptions can be
excepted at runtime.