Remove assertion about bare raises in raises visitor.
The Django repository apparently has a function which is intended
to be used in the context of an exception handler.  It seems like
bad style to me, but if it can occur in the wild, we probably don't
want to fail an assertion on it.  (Although, that will only happen
during tests or while debugging.)